bot/requirements.txt
requests~=2.31.0
telebot~=0.0.5
python-dotenv~=1.0.1
aiohttp~=3.9.5
telethon~=1.35.0